# openapi.yaml
openapi: 3.0.3
info:
  title: Placement Hub API
  version: "1.0"
  description: |
    Student, recruiter and placement-cell API. Auth is a short-lived Bearer access token; the refresh token
    lives in an httpOnly cookie (`ph_refresh`, path `/auth`). Errors always look like
    `{"error": {"code": "...", "message": "...", "fields": {"field": "reason"}}}`.

    A student who accepts an offer is PLACED: applying to jobs or registering for drives then fails with
    `403 already_placed`, while their history stays readable.
servers:
  - url: http://localhost:8080
security:
  - bearer: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: 'Sign in (sets the refresh cookie). Rate limited.'
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/LoginRequest"}}}}
      responses:
        "200": {description: Session, content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: 'Rotate the refresh cookie and get a new access token. Reusing an old cookie revokes the session.'
      responses:
        "200": {description: Session, content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}}
        "401": {$ref: "#/components/responses/Error"}
...
  /auth/register/company:
    post:
      tags: [Auth]
      security: []
      summary: 'Register a recruiter account. Starts Pending; may draft jobs but not submit them.'
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CompanyRegistration"}}}}
      responses:
        "201": {description: Registered}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /auth/change-password:
    post:
      tags: [Auth]
      summary: 'Change password (required after a temporary one). Signs out other sessions and returns a new session.'
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [currentPassword, newPassword], properties: {currentPassword: {type: string}, newPassword: {type: string, minLength: 8, maxLength: 72}}}}}}
      responses:
        "200": {description: Session, content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}}
        "422": {$ref: "#/components/responses/Error"}
...
  /jobs:
    get:
      tags: [Jobs]
      summary: 'Open jobs with per-student eligibility (`eligible`, `applied`, `canApply`, `blockReason`)'
      parameters:
        - {name: type, in: query, schema: {type: string, enum: [All, Full-time, Internship]}}
        - {name: q, in: query, schema: {type: string}}
        - {name: eligible, in: query, schema: {type: boolean}}
        - {name: sort, in: query, schema: {type: string, enum: [deadline, ctc]}}
        - {$ref: "#/components/parameters/page"}
        - {$ref: "#/components/parameters/limit"}
      responses: {"200": {description: Page of jobs, content: {application/json: {schema: {$ref: "#/components/schemas/JobPage"}}}}}
...
  /drives:
    get:
      tags: [Student]
      summary: 'Upcoming/ongoing drives (`scope=past` for finished) with `isRegistered`, `canRegister`, `blockReason`'
      parameters: [{name: scope, in: query, schema: {type: string, enum: [upcoming, past]}}]
      responses: {"200": {description: Items}}
  /drives/{id}/register:
    post: {tags: [Student], summary: "Register (idempotent). 403 already_placed; 422 when ineligible or the drive has started.", parameters: [{$ref: "#/components/parameters/id"}], responses: {"200": {description: Drive}}}
    delete: {tags: [Student], summary: "Cancel a registration before the drive starts", parameters: [{$ref: "#/components/parameters/id"}], responses: {"204": {description: Cancelled}}}

  # ---- notifications (students) ----------------------------------------------
...
  /company/jobs:
    get: {tags: [Company], summary: "All own postings in every status", responses: {"200": {description: Items}}}
    post:
      tags: [Company]
      summary: 'Create a draft; `"submit": true` also sends it for approval'
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/JobInput"}}}}
      responses: {"201": {description: Job}, "422": {$ref: "#/components/responses/Error"}}
...
  /admin/students:
    get:
      tags: [Admin]
      summary: 'Students with derived status (Placed / In process / Unplaced)'
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: branch, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string, enum: [All, Placed, In process, Unplaced]}}
        - {$ref: "#/components/parameters/page"}
        - {$ref: "#/components/parameters/limit"}
      responses: {"200": {description: Page}}
    post: {tags: [Admin], summary: "Create a student; returns a one-time temporary password", responses: {"201": {description: Created}}}
...
  /admin/companies:
    get: {tags: [Admin], parameters: [{name: status, in: query, schema: {type: string}}], summary: "Companies with derived openRoles, hires, lastVisit", responses: {"200": {description: Items}}}
...
  /admin/jobs:
    get: {tags: [Admin], parameters: [{name: status, in: query, schema: {type: string, enum: [All, Pending, Open, Closed, Rejected]}}], summary: "Submitted jobs (drafts are private)", responses: {"200": {description: Items}}}
...
  /admin/drives:
    get: {tags: [Admin], parameters: [{name: status, in: query, schema: {type: string, enum: [All, Upcoming, Ongoing, Completed]}}], summary: "Drives with registered/eligible counts", responses: {"200": {description: Items}}}
    post: {tags: [Admin], summary: "Schedule a drive; eligible students are notified", responses: {"201": {description: Drive}}}
  /admin/drives/{id}:
    get: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "One drive", responses: {"200": {description: Drive}}}
    put: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Edit; registered students are notified only if time, mode or venue change", responses: {"200": {description: Drive}}}
    delete: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Cancel (registered students are told)", responses: {"204": {description: Deleted}}}
  /admin/drives/{id}/registrations:
    get: {tags: [Admin], parameters: [{$ref: "#/components/parameters/id"}], summary: "Roster", responses: {"200": {description: Items}}}
...
  /notices:
    get: {tags: [Public], security: [], summary: "Notice board", parameters: [{$ref: "#/components/parameters/limit"}], responses: {"200": {description: "{items: [{id, date, tag, title}]}"}}}
...
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  responses:
    Error:
      description: Error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: {type: string, example: already_placed}
            message: {type: string}
            fields: {type: object, additionalProperties: {type: string}}
    LoginRequest: {type: object, required: [email, password], properties: {email: {type: string}, password: {type: string}}}
    CompanyRegistration:
      type: object
      required: [companyName, hrName, email, password]
      properties: {companyName: {type: string}, industry: {type: string}, hrName: {type: string}, email: {type: string}, phone: {type: string}, password: {type: string}}
    Session:
      type: object
      properties:
        accessToken: {type: string}
        expiresAt: {type: string, format: date-time}
        user: {type: object, properties: {id: {type: string, format: uuid}, email: {type: string}, name: {type: string}, role: {type: string, enum: [student, company, admin]}, mustChangePassword: {type: boolean}}}
...
    Job:
      type: object
      properties:
        id: {type: string}
        companyId: {type: string}
        company: {type: string}
        color: {type: string}
        role: {type: string}
        type: {type: string, enum: [Full-time, Internship]}
        location: {type: string}
        ctc: {type: number}
        minCgpa: {type: number}
        branches: {type: array, items: {type: string}}
        skills: {type: array, items: {type: string}}
        deadline: {type: string, format: date}
        openings: {type: integer}
        applicants: {type: integer}
        status: {type: string, enum: [Draft, Pending, Open, Closed, Rejected]}
        description: {type: string}
        allowBacklogs: {type: boolean}
        rejectReason: {type: string}
    StudentJob:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            eligible: {type: boolean}
            applied: {type: boolean}
            applicationId: {type: string, nullable: true}
            canApply: {type: boolean}
            blockReason: {type: object, nullable: true, properties: {code: {type: string}, message: {type: string}}}
    JobPage:
      type: object
      properties: {items: {type: array, items: {$ref: "#/components/schemas/StudentJob"}}, total: {type: integer}, page: {type: integer}, limit: {type: integer}}
    JobInput:
      type: object
      required: [role, type, location, ctc, branches, deadline, openings, description]
      properties:
        role: {type: string}
        type: {type: string, enum: [Full-time, Internship]}
        location: {type: string}
        ctc: {type: number}
        minCgpa: {type: number}
        branches: {type: array, items: {type: string}}
        skills: {type: array, items: {type: string}}
        deadline: {type: string, format: date}
        openings: {type: integer}
        description: {type: string}
        allowBacklogs: {type: boolean}
        submit: {type: boolean}
...